Show Progress Spinner when loading Bookmarks and Downloaded Videos tabs. Merged Bookmarks & Downloaded Grid Adapters into OrderableGridAdapter, and created new VideosGridFragment sub-class that supports OrderableGridAdapter.
