+++ app/src/main/java/free/rm/skytube/businessobjects/interfaces/OrderableDatabase.java
@@ -1,0 +1,12 @@
+package free.rm.skytube.businessobjects.interfaces;
+
+import java.util.List;
+
+import free.rm.skytube.businessobjects.YouTubeVideo;
+
+/**
+ * An interface to be used by a SQLiteOpenHelperEx database object, indicating that the videos in the database may be reordered.
+ */
+public interface OrderableDatabase {
+	void updateOrder(List<YouTubeVideo> videos);
+}
