+++ app/src/main/java/free/rm/skytube/gui/fragments/DownloadedVideosFragment.java
@@ -2,10 +2,7 @@ package free.rm.skytube.gui.fragments;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v7.widget.helper.ItemTouchHelper;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 
 import butterknife.BindView;
 import free.rm.skytube.R;
@@ -13,40 +10,18 @@ import free.rm.skytube.app.SkyTubeApp;
 import free.rm.skytube.businessobjects.AsyncTaskParallel;
 import free.rm.skytube.businessobjects.VideoCategory;
 import free.rm.skytube.businessobjects.db.DownloadedVideosDb;
-import free.rm.skytube.gui.businessobjects.DownloadedVideoGridAdapter;
-import free.rm.skytube.gui.businessobjects.SimpleItemTouchHelperCallback;
+import free.rm.skytube.gui.businessobjects.OrderableVideoGridAdapter;
+import free.rm.skytube.gui.businessobjects.fragments.OrderableVideosGridFragment;
 
-public class DownloadedVideosFragment extends VideosGridFragment implements DownloadedVideosDb.DownloadedVideosListener {
-	private DownloadedVideoGridAdapter downloadedVideoGridAdapter;
-
+public class DownloadedVideosFragment extends OrderableVideosGridFragment implements DownloadedVideosDb.DownloadedVideosListener {
 	@BindView(R.id.noDownloadedVideosText)
 	View noDownloadedVideosText;
 
 	@Override
 	public void onCreate(@Nullable Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
+		videoGridAdapter = new OrderableVideoGridAdapter(getActivity(), DownloadedVideosDb.getVideoDownloadsDb());
 		setLayoutResource(R.layout.videos_gridview_downloads);
-	}
-
-	@Override
-	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-		View view = super.onCreateView(inflater, container, savedInstanceState);
-
-		videoGridAdapter = null;
-
-		if(downloadedVideoGridAdapter ==  null) {
-			downloadedVideoGridAdapter = new DownloadedVideoGridAdapter(getActivity(), this);
-		} else {
-			downloadedVideoGridAdapter.setContext(getActivity());
-		}
-
-		gridView.setAdapter(downloadedVideoGridAdapter);
-
-		ItemTouchHelper.Callback callback = new SimpleItemTouchHelperCallback(downloadedVideoGridAdapter);
-		ItemTouchHelper touchHelper = new ItemTouchHelper(callback);
-		touchHelper.attachToRecyclerView(gridView);
-
-		return view;
 	}
 
 	@Override
@@ -79,7 +54,7 @@ public class DownloadedVideosFragment extends VideosGridFragment implements Down
 	@Override
 	public void onDownloadedVideosUpdated() {
 		populateList();
-		downloadedVideoGridAdapter.refresh();
+		videoGridAdapter.refresh();
 	}
 
 	private void populateList() {
@@ -112,19 +87,9 @@ public class DownloadedVideosFragment extends VideosGridFragment implements Down
 				noDownloadedVideosText.setVisibility(View.GONE);
 
 				// set video category and get the bookmarked videos asynchronously
-				downloadedVideoGridAdapter.setVideoCategory(VideoCategory.DOWNLOADED_VIDEOS);
+				videoGridAdapter.setVideoCategory(VideoCategory.DOWNLOADED_VIDEOS);
 			}
 		}
 
-	}
-
-	@Override
-	public void onRefresh() {
-		downloadedVideoGridAdapter.refresh(new Runnable() {
-			@Override
-			public void run() {
-				swipeRefreshLayout.setRefreshing(false);
-			}
-		});
 	}
 }
